moving rest to configdefaults
